betabinexch(LearnBayes)	R Documentation

Log posterior of logit mean and log precision for Binomial/beta exchangeable model

Description

Computes the log posterior density of logit mean and log precision for a Binomial/beta exchangeable model

Usage


betabinexch(theta,data)

Arguments

theta 	 vector of parameter values of logit eta and log K

data 	 a matrix with columns y (counts) and n (sample sizes)

Value

value of the log posterior

Author(s)

Jim Albert

Examples


n=c(20,20,20,20,20)
y=c(1,4,3,6,10)
data=cbind(y,n)
theta=c(-1,0)
betabinexch(theta,data)

[Package LearnBayes version 2.0 Index]