bfexch(LearnBayes)	R Documentation

Logarithm of integral of Bayes factor for testing homogeneity of proportions

Description

Computes the logarithm of the integral of the Bayes factor for testing homogeneity of a set of proportions

Usage


bfexch(theta,datapar)

Arguments

theta 	 value of the logit of the prior mean hyperparameter

datapar 	 list with components data, matrix with columns y (counts) and n (sample sizes), and K, prior precision hyperparameter

Value

value of the logarithm of the integral

Author(s)

Jim Albert

Examples


y=c(1,3,2,4,6,4,3)
n=c(10,10,10,10,10,10,10)
data=cbind(y,n)
K=20
datapar=list(data=data,K=K)
theta=1
bfexch(theta,datapar)

[Package LearnBayes version 2.0 Index]