Computes the posterior for binomial sampling and a mixture of betas prior

Computes the parameters and mixing probabilities for a binomial sampling problem where the prior is a discrete mixture of beta densities.

binomial.beta.mix(probs,betapar,data)

Arguments

probs
vector of probabilities of the beta components of the prior

betapar
matrix where each row contains the shape parameters for a beta component of the prior

data
vector of number of successes and number of failures

Value

probs

vector of probabilities of the beta components of the posterior

betapar

matrix where each row contains the shape parameters for a beta component of the posterior

Examples

probs=c(.5, .5)
beta.par1=c(15,5)
beta.par2=c(10,10)
betapar=rbind(beta.par1,beta.par2)
data=c(20,15)
binomial.beta.mix(probs,betapar,data)

#> $probs
#> beta.par1 beta.par2 
#> 0.2917961 0.7082039 
#> 
#> $betapar
#>           [,1] [,2]
#> beta.par1   35   20
#> beta.par2   30   25
#>