birthweight(LearnBayes)	R Documentation

Birthweight regression study

Description

Dobson describes a study where one is interested in predicting a baby's birthweight based on the gestational age and the baby's gender.

Usage


birthweight

Format

A data frame with 24 observations on the following 3 variables.

age

gestational age in weeks

gender

gender of the baby where 0 (1) is male (female)

weight

birthweight of baby in grams

Source

Dobson, A. (2001), An Introduction to Generalized Linear Models, New York: Chapman and Hall.

[Package LearnBayes version 2.0 Index]