Simulates values of predicted response for linear regression model

Simulates draws of the predictive distribution of a future response for a linear regression model with a noninformative prior

blinregpred(X1,theta.sample)

Arguments

X1
matrix where each row corresponds to a covariate set

theta.sample
list with components beta, matrix of simulated draws of regression vector, and sigma, vector of simulated draws of sampling error standard deviation

Value

matrix where a column corresponds to the simulated draws of the predicted response for a given covariate set

Examples

chirps=c(20,16.0,19.8,18.4,17.1,15.5,14.7,17.1,15.4,16.2,15,17.2,16,17,14.1)
temp=c(88.6,71.6,93.3,84.3,80.6,75.2,69.7,82,69.4,83.3,78.6,82.6,80.6,83.5,76.3)
X=cbind(1,chirps)
m=1000
theta.sample=blinreg(temp,X,m)
covset1=c(1,15)
covset2=c(1,20)
X1=rbind(covset1,covset2)
blinregpred(X1,theta.sample)

#>             [,1]      [,2]
#>    [1,] 72.44477  87.90996
#>    [2,] 76.02943  83.40029
#>    [3,] 70.85773  88.33767
#>    [4,] 73.95606  93.05931
#>    [5,] 78.46425  89.67294
#>    [6,] 74.22952  90.99475
#>    [7,] 74.24675  88.32839
#>    [8,] 67.64114  85.91519
#>    [9,] 71.29115  95.70305
#>   [10,] 73.62975  85.04344
#>   [11,] 79.68284  82.84064
#>   [12,] 74.56157  91.83569
#>   [13,] 68.13018 102.04031
#>   [14,] 76.96578  75.68815
#>   [15,] 71.44049  95.52532
#>   [16,] 70.00642  92.79930
#>   [17,] 79.20526  99.67471
#>   [18,] 77.75574  89.92660
#>   [19,] 87.51179  86.99850
#>   [20,] 75.25089  87.45507
#>   [21,] 73.79076  89.15377
#>   [22,] 78.43243  89.84193
#>   [23,] 77.15709  96.74859
#>   [24,] 73.59859  89.45215
#>   [25,] 66.27952 101.39317
#>   [26,] 73.62818  91.26815
#>   [27,] 79.16475  96.19175
#>   [28,] 74.90499  88.02357
#>   [29,] 77.81404  85.06244
#>   [30,] 75.90998  89.58765
#>   [31,] 74.61107  86.73516
#>   [32,] 74.42100  93.97462
#>   [33,] 66.81820  94.93751
#>   [34,] 81.69983  93.63807
#>   [35,] 81.85665  81.99996
#>   [36,] 76.14227  91.07119
#>   [37,] 74.79131  89.15400
#>   [38,] 82.06689  81.91600
#>   [39,] 82.57743  93.47951
#>   [40,] 75.62600  91.37159
#>   [41,] 75.41062  80.72320
#>   [42,] 79.14639  93.20307
#>   [43,] 66.57854  96.08205
#>   [44,] 75.78025  94.68382
#>   [45,] 70.22393  91.42661
#>   [46,] 71.24706  97.17712
#>   [47,] 76.96951  89.21402
#>   [48,] 75.56432  94.90691
#>   [49,] 78.23721  92.32517
#>   [50,] 74.16700  87.90156
#>   [51,] 73.55230  99.68831
#>   [52,] 73.23374  98.90491
#>   [53,] 78.71220  85.12360
#>   [54,] 75.50642  93.22871
#>   [55,] 67.66514 101.58830
#>   [56,] 72.39845  85.38354
#>   [57,] 70.12182  84.14465
#>   [58,] 78.50791  88.82213
#>   [59,] 72.58636  89.55181
#>   [60,] 77.54054  89.22475
#>   [61,] 72.34491  92.64926
#>   [62,] 66.92307  82.74523
#>   [63,] 65.60731  94.59308
#>   [64,] 76.04045  97.53094
#>   [65,] 86.03950  95.27714
#>   [66,] 75.74459  92.25564
#>   [67,] 70.38562  91.87504
#>   [68,] 75.33146  91.92771
#>   [69,] 70.70097  90.80798
#>   [70,] 80.13012  93.87006
#>   [71,] 67.99876  91.42772
#>   [72,] 71.15435  87.86630
#>   [73,] 74.92231  92.93067
#>   [74,] 70.60465  95.24082
#>   [75,] 83.13265  90.73617
#>   [76,] 68.07768  89.50533
#>   [77,] 75.98637  78.44875
#>   [78,] 72.88848  93.50731
#>   [79,] 76.99903  95.82373
#>   [80,] 78.96715  84.50813
#>   [81,] 74.03926  84.94118
#>   [82,] 69.48375  94.72317
#>   [83,] 78.59846  87.46715
#>   [84,] 83.94911  85.75655
#>   [85,] 76.43939  96.97847
#>   [86,] 78.48983  91.16650
#>   [87,] 78.65070  85.27549
#>   [88,] 74.32146  90.74311
#>   [89,] 75.56274  88.15920
#>   [90,] 79.71776  85.32412
#>   [91,] 76.69503  88.90352
#>   [92,] 79.72546  92.53534
#>   [93,] 75.52482  88.55981
#>   [94,] 75.65601  88.20606
#>   [95,] 71.78899  91.77157
#>   [96,] 72.06077  91.91874
#>   [97,] 73.06153  93.99722
#>   [98,] 69.55412  91.08964
#>   [99,] 78.63221  85.90889
#>  [100,] 68.99537  95.14314
#>  [101,] 75.13957  88.16757
#>  [102,] 65.46054  86.85817
#>  [103,] 70.74140  79.86804
#>  [104,] 74.49526  86.69691
#>  [105,] 72.19878  98.57515
#>  [106,] 75.32526  85.35128
#>  [107,] 72.86322  85.34266
#>  [108,] 79.96966  98.03964
#>  [109,] 79.34675  90.10191
#>  [110,] 82.51427  87.51285
#>  [111,] 59.43742  93.58317
#>  [112,] 70.09016  89.71484
#>  [113,] 76.60307  86.58749
#>  [114,] 74.31546  99.84368
#>  [115,] 70.75113  89.84455
#>  [116,] 71.58581  90.60618
#>  [117,] 66.74207  92.67321
#>  [118,] 74.51925  92.21138
#>  [119,] 69.85963  91.63885
#>  [120,] 74.88182  94.59569
#>  [121,] 78.98322  92.09811
#>  [122,] 71.14479  90.45555
#>  [123,] 67.41812  92.12299
#>  [124,] 72.71978  86.04533
#>  [125,] 67.14605  99.78643
#>  [126,] 68.48310 103.06183
#>  [127,] 71.71913  95.00003
#>  [128,] 69.91195  93.80846
#>  [129,] 72.28687  89.48134
#>  [130,] 91.51941  86.47813
#>  [131,] 70.18675  90.14637
#>  [132,] 68.92493  90.22792
#>  [133,] 72.94478  94.21527
#>  [134,] 71.15697  83.23744
#>  [135,] 71.74677  90.48897
#>  [136,] 74.91778  88.32132
#>  [137,] 73.69487  88.52527
#>  [138,] 73.27325  95.61345
#>  [139,] 71.79592  95.95358
#>  [140,] 79.29386  93.35993
#>  [141,] 76.39948  95.80358
#>  [142,] 72.50183  89.23223
#>  [143,] 72.44515  94.56005
#>  [144,] 79.08814  97.45153
#>  [145,] 75.58825  91.19187
#>  [146,] 69.24170  96.51534
#>  [147,] 78.04631  91.59798
#>  [148,] 73.90671  84.48572
#>  [149,] 84.23440  93.95588
#>  [150,] 80.26618  92.23459
#>  [151,] 76.32922  86.55666
#>  [152,] 72.59572  87.67657
#>  [153,] 78.09565  89.07661
#>  [154,] 76.30789  83.14099
#>  [155,] 67.77706  90.64078
#>  [156,] 74.65596  88.94711
#>  [157,] 83.36457  89.75924
#>  [158,] 78.64153  85.73858
#>  [159,] 69.25293  85.73064
#>  [160,] 73.19550  94.72160
#>  [161,] 73.25069  88.92771
#>  [162,] 72.98932  94.34438
#>  [163,] 74.80632  93.84989
#>  [164,] 72.26496  91.72938
#>  [165,] 80.39116  89.44885
#>  [166,] 76.41449  93.13368
#>  [167,] 78.27584  90.00133
#>  [168,] 71.31564  90.38868
#>  [169,] 70.08568  86.95386
#>  [170,] 74.47487  81.76717
#>  [171,] 73.33565  93.99818
#>  [172,] 76.99321  96.98415
#>  [173,] 75.23499  83.66900
#>  [174,] 73.88548  80.32283
#>  [175,] 70.64462  89.02877
#>  [176,] 75.12799  88.70037
#>  [177,] 75.04575  90.77900
#>  [178,] 67.28190  92.40938
#>  [179,] 76.90679  82.78597
#>  [180,] 75.74138  96.67033
#>  [181,] 70.29167  95.00501
#>  [182,] 83.15446  88.04526
#>  [183,] 75.67927  91.16266
#>  [184,] 75.45463  88.92743
#>  [185,] 81.62865  96.43400
#>  [186,] 74.32210  98.69742
#>  [187,] 79.70626  89.51279
#>  [188,] 80.81036  94.18070
#>  [189,] 72.96372  98.10533
#>  [190,] 77.20309  92.53151
#>  [191,] 80.34279  94.98056
#>  [192,] 71.78211  84.81143
#>  [193,] 75.57617  95.16769
#>  [194,] 78.28865  85.14062
#>  [195,] 73.52788  87.99792
#>  [196,] 74.64611  89.34032
#>  [197,] 68.06379  89.10918
#>  [198,] 69.93327  88.06322
#>  [199,] 63.82838  90.35100
#>  [200,] 67.98191  85.82066
#>  [201,] 76.58097  88.95592
#>  [202,] 69.50121  98.19965
#>  [203,] 74.86236  92.97518
#>  [204,] 70.17479  89.72238
#>  [205,] 72.66525  87.62005
#>  [206,] 74.09087  84.04679
#>  [207,] 79.66956  95.08067
#>  [208,] 75.53354  95.68988
#>  [209,] 77.36011  90.24757
#>  [210,] 68.66484  87.63081
#>  [211,] 71.40240  92.00995
#>  [212,] 80.23903  89.17962
#>  [213,] 75.18521  86.06526
#>  [214,] 79.09573  97.71244
#>  [215,] 66.82935  91.70624
#>  [216,] 74.32581  89.03565
#>  [217,] 77.66950  99.73984
#>  [218,] 78.07804  92.72059
#>  [219,] 69.75999  91.67382
#>  [220,] 81.25369  89.86451
#>  [221,] 71.18266  87.80100
#>  [222,] 70.39685  94.24052
#>  [223,] 76.02244  90.01830
#>  [224,] 67.88941  94.11913
#>  [225,] 79.89553  96.20632
#>  [226,] 75.30190  91.63619
#>  [227,] 72.22171 100.11252
#>  [228,] 68.89251  95.38126
#>  [229,] 73.89250  94.74608
#>  [230,] 73.65087  92.27910
#>  [231,] 77.04899  99.64259
#>  [232,] 74.92464  94.09757
#>  [233,] 71.61327  85.50684
#>  [234,] 86.65633  90.10015
#>  [235,] 71.02451  87.03062
#>  [236,] 71.18850  92.31304
#>  [237,] 69.70590  83.32782
#>  [238,] 74.10149  88.81956
#>  [239,] 73.31806  88.09778
#>  [240,] 70.90497  91.20702
#>  [241,] 74.59764  93.88735
#>  [242,] 82.55374  92.37605
#>  [243,] 69.86183  97.00381
#>  [244,] 74.66966  88.64389
#>  [245,] 76.75652  92.61891
#>  [246,] 76.65925  88.85700
#>  [247,] 76.17038  90.51106
#>  [248,] 92.01389  83.42016
#>  [249,] 78.21211  88.07410
#>  [250,] 74.61115  88.90332
#>  [251,] 75.22845  86.66217
#>  [252,] 71.73626  94.69813
#>  [253,] 71.19884  93.56538
#>  [254,] 80.10792  88.41619
#>  [255,] 64.76969  95.85242
#>  [256,] 76.97034  92.28585
#>  [257,] 72.61549  77.06459
#>  [258,] 77.48197  85.56552
#>  [259,] 72.12690  89.95138
#>  [260,] 76.50596  87.84143
#>  [261,] 78.33577  88.72932
#>  [262,] 81.51045  89.25966
#>  [263,] 71.20585 100.83939
#>  [264,] 70.81453  89.16598
#>  [265,] 76.42781  89.39797
#>  [266,] 76.40826  91.92271
#>  [267,] 83.18538  91.04790
#>  [268,] 71.75036  89.90031
#>  [269,] 79.76709  96.64045
#>  [270,] 81.03465  86.91042
#>  [271,] 70.99056  96.05819
#>  [272,] 73.35773  89.25628
#>  [273,] 69.88618  90.35612
#>  [274,] 76.06101  89.42387
#>  [275,] 77.71818  91.29089
#>  [276,] 80.09451  92.75013
#>  [277,] 80.14774  89.55202
#>  [278,] 73.89905  89.45890
#>  [279,] 64.80393  98.95293
#>  [280,] 70.79883  96.35633
#>  [281,] 75.53393  97.84327
#>  [282,] 72.74787  97.18364
#>  [283,] 76.53482  90.60518
#>  [284,] 74.89912  98.76850
#>  [285,] 78.15960  86.18955
#>  [286,] 73.05644  94.82351
#>  [287,] 74.85590  91.86775
#>  [288,] 87.81029  82.43811
#>  [289,] 74.40315  88.42200
#>  [290,] 77.76429  89.69311
#>  [291,] 71.02007 102.72445
#>  [292,] 82.03970  79.29913
#>  [293,] 70.81624  94.30981
#>  [294,] 72.24742  91.40262
#>  [295,] 74.58879 102.59284
#>  [296,] 78.50952  84.26652
#>  [297,] 62.61672 101.69318
#>  [298,] 71.53329  86.42994
#>  [299,] 71.93440  93.62191
#>  [300,] 77.20240  89.31933
#>  [301,] 70.04968  91.79349
#>  [302,] 77.42383  97.30247
#>  [303,] 75.94289  86.69739
#>  [304,] 76.12057  95.29729
#>  [305,] 71.91524  90.44683
#>  [306,] 75.51393  97.59798
#>  [307,] 74.79217  83.94351
#>  [308,] 81.58389  78.96695
#>  [309,] 62.27171  91.61649
#>  [310,] 71.38802  88.13328
#>  [311,] 65.09200  88.06659
#>  [312,] 72.00596  86.87280
#>  [313,] 70.59053  91.85813
#>  [314,] 76.24101  90.41401
#>  [315,] 75.29265  85.05293
#>  [316,] 79.60221  99.26218
#>  [317,] 77.69623  96.02728
#>  [318,] 79.50756  89.41070
#>  [319,] 68.69303  89.22552
#>  [320,] 68.28381  93.24543
#>  [321,] 77.32703  84.81098
#>  [322,] 72.99388  96.25728
#>  [323,] 77.42804  90.33645
#>  [324,] 70.53585  92.18107
#>  [325,] 70.39970 103.72917
#>  [326,] 73.87245  87.32947
#>  [327,] 77.12730  95.73116
#>  [328,] 78.94711  90.34631
#>  [329,] 70.42059  93.37318
#>  [330,] 69.20240  96.11776
#>  [331,] 79.88236  87.89707
#>  [332,] 71.66237  84.64203
#>  [333,] 78.19110  95.27625
#>  [334,] 78.25458  86.16067
#>  [335,] 77.98044  93.36209
#>  [336,] 69.10191  81.49637
#>  [337,] 74.19088  93.45125
#>  [338,] 78.32081  88.22939
#>  [339,] 78.90730  84.49608
#>  [340,] 80.21481  94.91518
#>  [341,] 73.48065  96.44602
#>  [342,] 68.75174  88.38425
#>  [343,] 77.74027  91.63605
#>  [344,] 74.47693  94.37431
#>  [345,] 73.47207  94.89159
#>  [346,] 73.29507  81.45379
#>  [347,] 72.33803  93.51224
#>  [348,] 80.29740  91.22440
#>  [349,] 72.05393  96.83973
#>  [350,] 70.81246  90.42031
#>  [351,] 75.44850  87.94154
#>  [352,] 73.52482 102.36001
#>  [353,] 80.76751  95.05385
#>  [354,] 73.69078  89.10039
#>  [355,] 76.36789  95.07737
#>  [356,] 80.98411  88.74789
#>  [357,] 69.76306  97.09056
#>  [358,] 73.28838  85.26981
#>  [359,] 74.73287  86.98764
#>  [360,] 78.66437  75.18729
#>  [361,] 76.96556  96.36885
#>  [362,] 78.84121  81.04742
#>  [363,] 72.28760  93.27209
#>  [364,] 71.29418  88.35694
#>  [365,] 75.72041  90.80033
#>  [366,] 71.86754 101.01952
#>  [367,] 80.85379  89.76964
#>  [368,] 68.35520  88.71667
#>  [369,] 67.84805  86.00267
#>  [370,] 72.23061  91.24615
#>  [371,] 69.36275  86.98088
#>  [372,] 77.96739  84.90362
#>  [373,] 70.64582  91.22979
#>  [374,] 72.65360  96.74855
#>  [375,] 68.93971  92.37807
#>  [376,] 72.02747  94.44015
#>  [377,] 66.34624  88.77416
#>  [378,] 71.16484  83.28457
#>  [379,] 80.04744  85.56941
#>  [380,] 73.40301  86.58784
#>  [381,] 69.78152  92.68856
#>  [382,] 80.67324  82.00693
#>  [383,] 72.62537  91.98258
#>  [384,] 77.09148  87.37224
#>  [385,] 63.97886  85.91158
#>  [386,] 66.93278  91.06392
#>  [387,] 76.40713  98.51675
#>  [388,] 70.67524  94.38025
#>  [389,] 70.66655  88.67134
#>  [390,] 70.24488  95.79143
#>  [391,] 74.61207  85.09848
#>  [392,] 79.13576  90.18442
#>  [393,] 80.34797  93.88757
#>  [394,] 70.11295  94.20739
#>  [395,] 75.69922  87.01483
#>  [396,] 73.39364  99.84991
#>  [397,] 72.24715  88.98968
#>  [398,] 86.60360  83.78236
#>  [399,] 73.62888  87.38150
#>  [400,] 73.72068  88.40083
#>  [401,] 79.96706  96.41728
#>  [402,] 73.92683  99.66974
#>  [403,] 76.20704  89.92090
#>  [404,] 72.53964  87.70533
#>  [405,] 77.54245  86.03291
#>  [406,] 69.11851  94.74240
#>  [407,] 81.08574  85.93691
#>  [408,] 81.58331  93.07031
#>  [409,] 76.29190  94.76811
#>  [410,] 75.80061  88.11426
#>  [411,] 75.36788  87.55828
#>  [412,] 73.53451  90.87427
#>  [413,] 77.42551  82.35112
#>  [414,] 71.95625  87.17716
#>  [415,] 78.09490  91.07297
#>  [416,] 75.94957  92.42263
#>  [417,] 72.87520  98.49678
#>  [418,] 72.13558  96.51042
#>  [419,] 76.78166  92.08976
#>  [420,] 72.52540  96.10236
#>  [421,] 74.58604  92.26122
#>  [422,] 72.06058  85.27431
#>  [423,] 73.46569  98.63006
#>  [424,] 73.30944  93.38596
#>  [425,] 71.91649  96.13331
#>  [426,] 71.26281  87.97012
#>  [427,] 74.26582  88.01020
#>  [428,] 76.60989  82.15740
#>  [429,] 80.59463  92.29210
#>  [430,] 80.63884  90.19619
#>  [431,] 70.84732  97.83360
#>  [432,] 77.36134  95.75804
#>  [433,] 74.93805  88.77614
#>  [434,] 78.21013 100.72895
#>  [435,] 75.31817  94.49315
#>  [436,] 76.89690  88.82320
#>  [437,] 80.01358  88.60889
#>  [438,] 78.11152  93.84755
#>  [439,] 73.63655  90.83411
#>  [440,] 73.91833  99.92075
#>  [441,] 74.77734  92.49226
#>  [442,] 72.95567  83.88248
#>  [443,] 77.21695  85.18528
#>  [444,] 74.60735  85.48140
#>  [445,] 74.32745  90.15517
#>  [446,] 69.73268  95.39035
#>  [447,] 79.39670  93.27579
#>  [448,] 72.30321  75.25898
#>  [449,] 68.08702  89.98916
#>  [450,] 74.56986  87.58039
#>  [451,] 69.04646  85.33903
#>  [452,] 78.61309  95.17799
#>  [453,] 81.78389  91.82923
#>  [454,] 81.94559  87.96516
#>  [455,] 74.55573  79.98682
#>  [456,] 80.33353  87.31542
#>  [457,] 68.81295  88.80062
#>  [458,] 72.91949  93.86397
#>  [459,] 78.56371  90.23441
#>  [460,] 75.06676  92.73503
#>  [461,] 77.07615  99.47535
#>  [462,] 71.03942  88.65086
#>  [463,] 68.57141  90.05049
#>  [464,] 75.68869  91.74932
#>  [465,] 75.72285  80.11680
#>  [466,] 69.81215  99.69833
#>  [467,] 67.87725  85.07632
#>  [468,] 72.88473  85.86466
#>  [469,] 72.68582  93.79045
#>  [470,] 75.62930  90.16120
#>  [471,] 73.47095  98.60588
#>  [472,] 78.35472  88.03199
#>  [473,] 77.74977  92.69929
#>  [474,] 78.65095  81.04105
#>  [475,] 75.92143  85.65387
#>  [476,] 70.13274  97.04288
#>  [477,] 70.87493  92.72811
#>  [478,] 78.56835  89.97237
#>  [479,] 72.03860  97.52751
#>  [480,] 71.44983  91.17813
#>  [481,] 68.81428  85.79479
#>  [482,] 83.64601  91.31959
#>  [483,] 67.01190  90.08173
#>  [484,] 74.38818  97.92804
#>  [485,] 74.58294  86.83960
#>  [486,] 76.48775  93.96815
#>  [487,] 86.85882  80.83397
#>  [488,] 69.80929  92.61509
#>  [489,] 78.08560  89.07288
#>  [490,] 79.40028  89.64878
#>  [491,] 79.09620  95.55399
#>  [492,] 67.82180  86.85220
#>  [493,] 70.03164  88.17205
#>  [494,] 66.29771  97.31045
#>  [495,] 80.69144  89.73662
#>  [496,] 69.17861  92.15923
#>  [497,] 78.00742  88.41798
#>  [498,] 73.06661  87.61095
#>  [499,] 81.41501  90.26180
#>  [500,] 79.67592  90.55284
#>  [ reached getOption("max.print") -- omitted 500 rows ]