bradley.terry.post {LearnBayes}	R Documentation

Log posterior of a Bradley Terry random effects model

Description

Computes the log posterior density of the talent parameters and the log standard deviation for a Bradley Terry model with normal random effects

Usage

bradley.terry.post(theta,data)

Arguments

theta
vector of talent parameters and log standard deviation

data
data matrix with columns team1, team2, wins by team1, and wins by team2

Value

value of the log posterior

Author(s)

Jim Albert

Examples

data(baseball.1964)
team.strengths=rep(0,10)
log.sigma=0
bradley.terry.post(c(team.strengths,log.sigma),baseball.1964)

[Package LearnBayes version 2.15.1 Index]