calculus.grades(LearnBayes)	R Documentation

Calculus grades dataset

Description

Grades and other variables collected for a sample of calculus students.

Usage


calculus.grades

Format

A data frame with 100 observations on the following 3 variables.

grade

indicates if student received a A or B in class

prev.grade

indicates if student received a A in prerequisite math class

act

score on the ACT math test

Source

Collected by a colleague of the author at his university.

[Package LearnBayes version 2.0 Index]