cancermortality(LearnBayes)	R Documentation

Cancer mortality data

Description

Number of cancer deaths and number at risk for 20 cities in Missouri.

Usage


cancermortality

Format

A data frame with 20 observations on the following 2 variables.

y

number of cancer deaths

n

number at risk

Source

Tsutakawa, R., Shoop, G., and Marienfeld, C. (1985), Empirical Bayes Estimation of Cancer Mortality Rates, Statistics in Medicine, 4, 201-212.

[Package LearnBayes version 2.0 Index]