careertraj.setup(LearnBayes)	R Documentation

Setup for Career Trajectory Application

Description

Setups the data matrices for the use of WinBUGS in the career trajectory application.

Usage


careertraj.setup(data)

Arguments

data 	 data matrix for ballplayers with variables Player, Year, Age, G, AB, R, H, X2B, X3B, HR, RBI, BB, SO

Value

player.names 	 vector of player names

y 	 matrix of home runs for players where a row corresponds to the home runs for a player during all the years of his career

n 	 matrix of AB-SO for all players

x 	 matrix of ages for all players for all years of their careers

T 	 vector of number of seasons for all players

N 	 number of players

Author(s)

Jim Albert

Examples


data(sluggerdata)
careertraj.setup(sluggerdata)

[Package LearnBayes version 2.0 Index]