discint(LearnBayes)	R Documentation

Highest probability interval for a discrete distribution

Description

Computes a highest probability interval for a discrete probability distribution

Usage


discint(dist, prob)

Arguments

dist 	 probability distribution written as a matrix where the first column contain the values and the second column the probabilities

prob 	 probability content of interest

Value

prob 	 exact probability content of interval

set 	 set of values of the probability interval

Author(s)

Jim Albert

Examples


x=0:10
probs=dbinom(x,size=10,prob=.3)
dist=cbind(x,probs)
pcontent=.8
discint(dist,pcontent)

[Package LearnBayes version 2.0 Index]