discrete.bayes {LearnBayes}	R Documentation

Posterior distribution with discrete priors

Description

Computes the posterior distribution for an arbitrary one parameter distribution for a discrete prior distribution.

Usage

discrete.bayes(df,prior,y,...)

Arguments

df
name of the function defining the sampling density

prior
vector defining the prior density; names of the vector define the parameter values and entries of the vector define the prior probabilities

y
vector of data values

...
any further fixed parameter values used in the sampling density function

Value

prob
vector of posterior probabilities

pred
scalar with prior predictive probability

Author(s)

Jim Albert

Examples

prior=c(.25,.25,.25,.25)
names(prior)=c(.2,.25,.3,.35)
y=5
n=10
discrete.bayes(dbinom,prior,y,size=n)

[Package LearnBayes version 2.15.1 Index]