discrete.bayes.2 {LearnBayes}	R Documentation

Posterior distribution of two parameters with discrete priors

Description

Computes the posterior distribution for an arbitrary two parameter distribution for a discrete prior distribution.

Usage

discrete.bayes.2(df,prior,y=NULL,...)

Arguments

df
name of the function defining the sampling density of two parameters

prior
matrix defining the prior density; the row names and column names of the matrix define respectively the values of parameter 1 and values of parameter 2 and the entries of the matrix give the prior probabilities

y
y is a matrix of data values, where each row corresponds to a single observation

...
any further fixed parameter values used in the sampling density function

Value

prob
matrix of posterior probabilities

pred
scalar with prior predictive probability

Author(s)

Jim Albert

Examples

p1 = seq(0.1, 0.9, length = 9)
p2 = p1
prior = matrix(1/81, 9, 9)
dimnames(prior)[[1]] = p1
dimnames(prior)[[2]] = p2
discrete.bayes.2(twoproplike,prior)

[Package LearnBayes version 2.15.1 Index]