dmnorm(LearnBayes)	R Documentation

The probability density function for the multivariate normal (Gaussian) probability distribution

Description

Computes the density of a multivariate normal distribution

Usage


dmnorm(x, mean = rep(0, d), varcov, log = FALSE)

Arguments

x 	 vector of length d or matrix with d columns, giving the coordinates of points where density is to evaluated

mean 	 numeric vector giving the location parameter of the distribution

varcov 	 a positive definite matrix representing the scale matrix of the distribution

log 	 a logical value; if TRUE, the logarithm of the density is to be computed

Value

vector of density values

Author(s)

Jim Albert

Examples


mu <- c(1,12,2)
Sigma <- matrix(c(1,2,0,2,5,0.5,0,0.5,3), 3, 3)
x <- c(2,14,0)
f <- dmnorm(x, mu, Sigma)

[Package LearnBayes version 2.0 Index]