donner(LearnBayes)	R Documentation

Donner survival study

Description

Data contains the age, gender and survival status for 45 members of the Donner Party who experienced difficulties in crossing the Sierra Nevada mountains in California.

Usage


donner

Format

A data frame with 45 observations on the following 3 variables.

age

age of person

male

gender that is 1 (0) if person is male (female)

survival

survival status, 1 or 0 if person survived or died

Source

Grayson, D. (1960), Donner party deaths: a demographic assessment, Journal of Anthropological Assessment, 46, 223-242.

[Package LearnBayes version 2.0 Index]