groupeddatapost(LearnBayes)	R Documentation

Log posterior of normal parameters when data is in grouped form

Description

Computes the log posterior density of (M,log S) for normal sampling where the data is observed in grouped form

Usage


groupeddatapost(theta,data)

Arguments

theta 	 vector of parameter values M and log S

data 	 list with components int.lo, a vector of left endpoints, int.hi, a vector of right endpoints, and f, a vector of bin frequencies

Value

value of the log posterior

Author(s)

Jim Albert

Examples


int.lo=c(-Inf,10,15,20,25)
int.hi=c(10,15,20,25,Inf)
f=c(2,5,8,4,2)
data=list(int.lo=int.lo,int.hi=int.hi,f=f)
theta=c(20,1)
groupeddatapost(theta,data)

[Package LearnBayes version 2.0 Index]