howardprior(LearnBayes)	R Documentation

Logarithm of Howard's dependent prior for two proportions

Description

Computes the logarithm of a dependent prior on two proportions proposed by Howard in a Statistical Science paper in 1998.

Usage


howardprior(xy,par)

Arguments

xy 	 vector of proportions p1 and p2

par 	 vector containing parameter values alpha, beta, gamma, delta, sigma

Value

value of the log posterior

Author(s)

Jim Albert

Examples


param=c(1,1,1,1,2)
p=c(.1,.5)
howardprior(p,param)

[Package LearnBayes version 2.0 Index]