lbinorm(LearnBayes)	R Documentation

Logarithm of bivariate normal density

Description

Computes the logarithm of a bivariate normal density

Usage


lbinorm(xy,par)

Arguments

xy 	 vector of values of two variables x and y

par 	 list with components m, a vector of means, and v, a variance-covariance matrix

Value

value of the kernel of the log density

Author(s)

Jim Albert

Examples


mean=c(0,0)
varcov=diag(c(1,1))
value=c(1,1)
param=list(m=mean,v=varcov)
lbinorm(value,param)

[Package LearnBayes version 2.0 Index]