logpoissnormal(LearnBayes)	R Documentation

Log posterior with Poisson sampling and normal prior

Description

Computes the logarithm of the posterior density of a Poisson log mean with a normal prior

Usage


logpoissnormal(theta,datapar)

Arguments

theta 	 vector of values of the log mean parameter

datapar 	 list with components data, vector of observations, and par, vector of parameters of the normal prior

Value

vector of values of the log posterior for all values in theta

Author(s)

Jim Albert

Examples


data=c(2,4,3,6,1,0,4,3,10,2)
par=c(0,1)
datapar=list(data=data,par=par)
theta=c(-1,0,1,2)
logpoissnormal(theta,datapar)

[Package LearnBayes version 2.0 Index]