mycontour(LearnBayes)	R Documentation

Contour plot of a bivariate density function

Description

For a general two parameter density, draws a contour graph where the contour lines are drawn at 10 percent, 1 percent, and .1 percent of the height at the mode.

Usage


mycontour(logf,limits,data,...)

Arguments

logf 	 function that defines the logarithm of the density

limits 	 limits (xlo, xhi, ylo, yhi) where the graph is to be drawn

data 	 vector or list of parameters associated with the function logpost

... 	 further arguments to pass to contour

Value

A contour graph of the density is drawn

Author(s)

Jim Albert

Examples


m=array(c(0,0),c(2,1))
v=array(c(1,.6,.6,1),c(2,2))
normpar=list(m=m,v=v)
mycontour(lbinorm,c(-4,4,-4,4),normpar)

[Package LearnBayes version 2.0 Index]