normal.normal.mix(LearnBayes)	R Documentation

Computes the posterior for normal sampling and a mixture of normals prior

Description

Computes the parameters and mixing probabilities for a normal sampling problem, variance known, where the prior is a discrete mixture of normal densities.

Usage


normal.normal.mix(probs,normalpar,data)

Arguments

probs 	 vector of probabilities of the normal components of the prior

normalpar 	 matrix where each row contains the mean and variance parameters for a normal component of the prior

data 	 vector of observation and sampling variance

Value

probs 	 vector of probabilities of the normal components of the posterior

normalpar 	 matrix where each row contains the mean and variance parameters for a normal component of the posterior

Author(s)

Jim Albert

Examples


probs=c(.5, .5)
normal.par1=c(0,1)
normal.par2=c(2,.5)
normalpar=rbind(normal.par1,normal.par2)
y=1; sigma2=.5
data=c(y,sigma2)
normal.normal.mix(probs,normalpar,data)

[Package LearnBayes version 2.0 Index]