normnormexch {LearnBayes}	R Documentation

Log posterior of mean and log standard deviation for Normal/Normal exchangeable model

Description

Computes the log posterior density of mean and log standard deviation for a Normal/Normal exchangeable model where (mean, log sd) is given a uniform prior.

Usage

normnormexch(theta,data)

Arguments

theta
vector of parameter values of mu and log tau

data
a matrix with columns y (observations) and v (sampling variances)

Value

value of the log posterior

Author(s)

Jim Albert

Examples

s.var <- c(0.05, 0.05, 0.05, 0.05, 0.05)
y.means <- c(1, 4, 3, 6,10)
data=cbind(y.means, s.var)
theta=c(-1, 0)
normnormexch(theta,data)

[Package LearnBayes version 2.15.1 Index]