normpostsim(LearnBayes)	R Documentation

Simulation from Bayesian normal sampling model

Description

Gives a simulated sample from the joint posterior distribution of the mean and variance for a normal sampling prior with a noninformative prior.

Usage


normpostsim(data,m=1000)

Arguments

data 	 vector of observations

m 	 number of simulations desired

Value

mu 	 vector of simulated draws of normal mean

sigma2 	 vector of simulated draws of normal variance

Author(s)

Jim Albert

Examples


data(darwin)
s=normpostsim(darwin$difference)

[Package LearnBayes version 2.0 Index]