ordergibbs(LearnBayes)	R Documentation

Gibbs sampling for a hierarchical regression model

Description

Implements Gibbs sampling for estimating a two-way table of means under a order restriction.

Usage


ordergibbs(data,m)

Arguments

data 	 data matrix with first two columns observed sample means and sample sizes

m 	 number of cycles of Gibbs sampling

Value

matrix of simulated draws of the normal means where each row represents one simulated draw

Author(s)

Jim Albert

Examples


data(iowagpa)
m=1000
s=ordergibbs(iowagpa,m)

[Package LearnBayes version 2.0 Index]