pbetap(LearnBayes)	R Documentation

Predictive distribution for a binomial sample with a beta prior

Description

Computes predictive distribution for number of successes of future binomial experiment with a discrete distribution for the proportion.

Usage


pbetap(ab, n, s)

Arguments

ab 	 vector of parameters of the beta prior

n 	 size of future binomial sample

s 	 vector of number of successes for future binomial experiment

Value

vector of predictive probabilities for the values in the vector s

Author(s)

Jim Albert

Examples


ab=c(3,12)
n=10
s=0:10
pbetap(ab,n,s)

[Package LearnBayes version 2.0 Index]