pdisc(LearnBayes)	R Documentation

Posterior distribution for a proportion with discrete priors

Description

Computes the posterior distribution for a proportion for a discrete prior distribution.

Usage


pdisc(p, prior, data)

Arguments

p 	 vector of proportion values

prior 	 vector of prior probabilities

data 	 vector consisting of number of successes and number of failures

Value

vector of posterior probabilities

Author(s)

Jim Albert

Examples


p=c(.2,.25,.3,.35)
prior=c(.25,.25,.25,.25)
data=c(5,10)
pdisc(p,prior,data)

[Package LearnBayes version 2.0 Index]