pdiscp(LearnBayes)	R Documentation

Predictive distribution for a binomial sample with a discrete prior

Description

Computes predictive distribution for number of successes of future binomial experiment with a discrete distribution for the proportion.

Usage


pdiscp(p, probs, n, s)

Arguments

p 	 vector of proportion values

probs 	 vector of probabilities

n 	 size of future binomial sample

s 	 vector of number of successes for future binomial experiment

Value

vector of predictive probabilities for the values in the vector s

Author(s)

Jim Albert

Examples


p=c(.1,.2,.3,.4,.5,.6,.7,.8,.9)
prob=c(0.05,0.10,0.10,0.15,0.20,0.15,0.10,0.10,0.05)
n=10
s=0:10
pdiscp(p,prob,n,s)

[Package LearnBayes version 2.0 Index]