poisson.gamma.mix(LearnBayes)	R Documentation

Computes the posterior for Poisson sampling and a mixture of gammas prior

Description

Computes the parameters and mixing probabilities for a Poisson sampling problem where the prior is a discrete mixture of gamma densities.

Usage


poisson.gamma.mix(probs,gammapar,data)

Arguments

probs 	 vector of probabilities of the gamma components of the prior

gammapar 	 matrix where each row contains the shape and rate parameters for a gamma component of the prior

data 	 list with components y, vector of counts, and t, vector of time intervals

Value

probs 	 vector of probabilities of the gamma components of the posterior

gammapar 	 matrix where each row contains the shape and rate parameters for a gamma component of the posterior

Author(s)

Jim Albert

Examples


probs=c(.5, .5)
gamma.par1=c(1,1)
gamma.par2=c(10,2)
gammapar=rbind(gamma.par1,gamma.par2)
y=c(1,3,2,4,10); t=c(1,1,1,1,1)
data=list(y=y,t=t)
poisson.gamma.mix(probs,gammapar,data)

[Package LearnBayes version 2.0 Index]