prior.two.parameters {LearnBayes}	R Documentation

Construct discrete uniform prior for two parameters

Description

Constructs a discrete uniform prior distribution for two parameters

Usage

prior.two.parameters(parameter1, parameter2)

Arguments

parameter1
vector of values of first parameter

parameter2
vector of values of second parameter

Value

matrix of uniform probabilities where the rows and columns are labelled with the parameter values

Author(s)

Jim Albert

Examples

prior.two.parameters(c(1,2,3,4),c(2,4,7))

[Package LearnBayes version 2.15.1 Index]