rmt {LearnBayes}	R Documentation

Random number generation for multivariate t

Description

Simulates from a multivariate t distribution

Usage

rmt(n = 1, mean = rep(0, d), S, df = Inf)

Arguments

n
number of random numbers to be generated

mean
numeric vector giving the location parameter of the distribution

S
a positive definite matrix representing the scale matrix of the distribution

df
degrees of freedom

Value

matrix of n rows of random vectors

Author(s)

Jim Albert

Examples

mu <- c(1,12,2)
Sigma <- matrix(c(1,2,0,2,5,0.5,0,0.5,3), 3, 3)
df <- 4
x <- rmt(10, mu, Sigma, df)

[Package LearnBayes version 2.15.1 Index]