strikeout(LearnBayes)	R Documentation

Baseball strikeout data

Description

For all professional baseball players in the 2004 season, dataset gives the number of strikeouts and at-bats when runners are in scoring position and when runners are not in scoring position.

Usage


strikeout

Format

A data frame with 438 observations on the following 4 variables.

r

number of strikeouts of player when runners are not in scoring position

n

number of at-bats of player when runners are not in scoring position

s

number of strikeouts of player when runners are in scoring position

m

number of at-bats of player when runners are in scoring position

Source

Collected from www.espn.com website.

[Package LearnBayes version 2.0 Index]