studentdata(LearnBayes)	R Documentation

Student dataset

Description

Answers to a sheet of questions given to a large number of students in introductory statistics classes

Usage


studentdata

Format

A data frame with 657 observations on the following 11 variables.

Student

student number

Height

height in inches

Gender

gender

Shoes

number of pairs of shoes owned

Number

number chosen between 1 and 10

Dvds

name of movie dvds owned

ToSleep

time the person went to sleep the previous night (hours past midnight)

WakeUp

time the person woke up the next morning

Haircut

cost of last haircut including tip

Job

number of hours working on a job per week

Drink

usual drink at suppertime among milk, water, and pop

Source

Collected by the author during the Fall 2006 semester.

[Package LearnBayes version 2.0 Index]