transplantpost(LearnBayes)	R Documentation

Log posterior of a Pareto model for survival data

Description

Computes the log posterior density of (log tau, log lambda, log p) for a Pareto model for survival data

Usage


transplantpost(theta,data)

Arguments

theta 	 vector of parameter values of log tau, log lambda, and log p

data 	 data matrix with columns survival time, transplant indicator, time to transplant, and censoring indicator

Value

value of the log posterior

Author(s)

Jim Albert

Examples


data(stanfordheart)
theta=c(0,3,-1)
transplantpost(theta,stanfordheart)

[Package LearnBayes version 2.0 Index]