weibullregpost {LearnBayes}	R Documentation

Log posterior of a Weibull proportional odds model for survival data

Description

Computes the log posterior density of (log sigma, mu, beta) for a Weibull proportional odds regression model

Usage

weibullregpost(theta,data)

Arguments

theta
vector of parameter values log sigma, mu, and beta

data
data matrix with columns survival time, censoring variable, and covariate matrix

Value

value of the log posterior

Author(s)

Jim Albert

Examples

data(chemotherapy)
attach(chemotherapy)
d=cbind(time,status,treat-1,age)
theta=c(-.6,11,.6,0)
weibullregpost(theta,d)

[Package LearnBayes version 2.15.1 Index]