bfindep {LearnBayes}	R Documentation

Bayes factor against independence assuming alternatives close to independence

Description

Computes a Bayes factor against independence for a two-way contingency table assuming a "close to independence" alternative model

Usage

bfindep(y,K,m)

Arguments

y
matrix of counts

K
Dirichlet precision hyperparameter

m
number of simulations

Value

bf
value of the Bayes factor against hypothesis of independence

nse
estimate of the simulation standard error of the computed Bayes factor

Author(s)

Jim Albert

Examples

y=matrix(c(10,4,6,3,6,10),c(2,3))
K=20
m=1000
bfindep(y,K,m)

[Package LearnBayes version 2.15.1 Index]