Simulates values of expected response for linear regression model

Simulates draws of the posterior distribution of an expected response for a linear regression model with a noninformative prior

blinregexpected(X1,theta.sample)

Arguments

X1
matrix where each row corresponds to a covariate set

theta.sample
list with components beta, matrix of simulated draws of regression vector, and sigma, vector of simulated draws of sampling error standard deviation

Value

matrix where a column corresponds to the simulated draws of the expected response for a given covariate set

Examples

chirps=c(20,16.0,19.8,18.4,17.1,15.5,14.7,17.1,15.4,16.2,15,17.2,16,17,14.1)
temp=c(88.6,71.6,93.3,84.3,80.6,75.2,69.7,82,69.4,83.3,78.6,82.6,80.6,83.5,76.3)
X=cbind(1,chirps)
m=1000
theta.sample=blinreg(temp,X,m)
covset1=c(1,15)
covset2=c(1,20)
X1=rbind(covset1,covset2)
blinregexpected(X1,theta.sample)

#>             [,1]     [,2]
#>    [1,] 74.62461 91.14482
#>    [2,] 75.06358 89.10516
#>    [3,] 75.39130 88.96936
#>    [4,] 74.06804 91.32665
#>    [5,] 72.96769 89.72732
#>    [6,] 75.16621 89.20575
#>    [7,] 74.70832 90.08955
#>    [8,] 74.51928 92.62399
#>    [9,] 76.32898 92.67834
#>   [10,] 74.33000 92.26559
#>   [11,] 74.20999 92.43889
#>   [12,] 74.87901 91.98854
#>   [13,] 73.92182 92.85846
#>   [14,] 75.48830 86.30887
#>   [15,] 75.31314 91.26558
#>   [16,] 75.96134 88.44334
#>   [17,] 74.09133 89.32397
#>   [18,] 75.53044 86.31802
#>   [19,] 74.76073 91.29746
#>   [20,] 73.31683 92.38463
#>   [21,] 73.33136 93.75395
#>   [22,] 74.95212 89.90963
#>   [23,] 75.10963 90.72766
#>   [24,] 76.85746 89.99015
#>   [25,] 75.94717 91.46236
#>   [26,] 74.17665 91.13413
#>   [27,] 74.43229 90.87550
#>   [28,] 73.44887 92.15665
#>   [29,] 75.53720 89.95846
#>   [30,] 77.16816 91.73981
#>   [31,] 73.83676 87.82945
#>   [32,] 75.35306 89.19534
#>   [33,] 74.42405 92.64298
#>   [34,] 70.98487 89.21642
#>   [35,] 72.29951 94.25886
#>   [36,] 70.23626 95.58513
#>   [37,] 74.98598 93.61519
#>   [38,] 75.74381 90.84292
#>   [39,] 75.29045 92.14613
#>   [40,] 74.86240 91.29922
#>   [41,] 76.52783 91.30118
#>   [42,] 71.54911 97.35234
#>   [43,] 74.92335 92.31504
#>   [44,] 77.58473 90.25687
#>   [45,] 75.57717 89.99972
#>   [46,] 74.09825 90.31151
#>   [47,] 74.37280 90.07561
#>   [48,] 72.50556 88.57770
#>   [49,] 76.96753 87.87956
#>   [50,] 74.73653 90.01920
#>   [51,] 76.11885 89.87766
#>   [52,] 74.91339 91.21022
#>   [53,] 73.99836 92.25794
#>   [54,] 72.61641 90.10448
#>   [55,] 73.49103 92.88220
#>   [56,] 77.78959 88.47882
#>   [57,] 73.05058 90.35343
#>   [58,] 75.01481 92.22415
#>   [59,] 76.77436 90.03420
#>   [60,] 76.86050 93.59088
#>   [61,] 76.42332 89.76565
#>   [62,] 76.17114 87.85865
#>   [63,] 73.43380 89.84760
#>   [64,] 76.79012 87.39829
#>   [65,] 77.49651 90.19903
#>   [66,] 74.20153 91.45003
#>   [67,] 74.59413 93.52522
#>   [68,] 75.22620 90.21059
#>   [69,] 72.64507 97.05308
#>   [70,] 78.33921 89.35404
#>   [71,] 73.48506 94.91282
#>   [72,] 75.76596 93.14965
#>   [73,] 74.22497 87.60212
#>   [74,] 75.43518 90.73758
#>   [75,] 74.52384 89.22688
#>   [76,] 73.18867 89.71424
#>   [77,] 73.94333 93.71141
#>   [78,] 73.13301 94.05051
#>   [79,] 73.38102 93.04938
#>   [80,] 74.66470 92.93863
#>   [81,] 73.93836 92.00413
#>   [82,] 73.58055 92.89712
#>   [83,] 75.14985 89.67034
#>   [84,] 74.71340 94.12858
#>   [85,] 73.71687 85.88361
#>   [86,] 75.57679 89.25478
#>   [87,] 73.62693 95.03275
#>   [88,] 75.60609 91.68778
#>   [89,] 74.25245 90.85767
#>   [90,] 75.78161 90.52689
#>   [91,] 73.95557 89.59464
#>   [92,] 74.58715 87.63063
#>   [93,] 72.18500 93.05363
#>   [94,] 72.71317 91.62825
#>   [95,] 75.30379 95.28383
#>   [96,] 75.51593 93.58871
#>   [97,] 74.09326 92.96785
#>   [98,] 77.97794 91.79528
#>   [99,] 77.39510 93.97801
#>  [100,] 74.25807 88.85620
#>  [101,] 75.64330 90.81495
#>  [102,] 71.29250 94.64964
#>  [103,] 74.13128 90.16473
#>  [104,] 73.52561 95.33454
#>  [105,] 75.42044 85.87168
#>  [106,] 75.09863 88.49594
#>  [107,] 72.26735 94.72256
#>  [108,] 72.50474 93.82552
#>  [109,] 75.51539 92.28857
#>  [110,] 71.95224 93.48607
#>  [111,] 74.91530 93.24648
#>  [112,] 73.71270 89.69272
#>  [113,] 73.53103 91.67274
#>  [114,] 73.94513 90.50369
#>  [115,] 75.20382 94.73721
#>  [116,] 73.96320 90.26517
#>  [117,] 75.79536 89.09713
#>  [118,] 77.60344 93.49585
#>  [119,] 71.97894 92.02966
#>  [120,] 75.60628 94.71495
#>  [121,] 77.86427 88.27012
#>  [122,] 74.72349 95.36084
#>  [123,] 70.68654 93.47021
#>  [124,] 75.95780 88.03128
#>  [125,] 73.95403 89.95329
#>  [126,] 74.30815 92.66028
#>  [127,] 73.93669 94.83941
#>  [128,] 75.95397 90.69623
#>  [129,] 74.17309 87.87463
#>  [130,] 78.86803 84.86762
#>  [131,] 76.33620 81.89649
#>  [132,] 76.05098 88.26287
#>  [133,] 74.36725 91.99727
#>  [134,] 76.02926 88.11848
#>  [135,] 71.56331 88.51768
#>  [136,] 73.32651 91.20213
#>  [137,] 75.37189 88.21177
#>  [138,] 73.76472 92.50759
#>  [139,] 76.91691 92.24357
#>  [140,] 72.77503 93.84059
#>  [141,] 73.53080 93.94694
#>  [142,] 74.90671 87.36366
#>  [143,] 73.19030 90.87489
#>  [144,] 75.49883 93.03042
#>  [145,] 76.15505 92.38821
#>  [146,] 74.13837 94.36964
#>  [147,] 73.20884 92.80249
#>  [148,] 73.25423 90.72852
#>  [149,] 75.80007 94.82449
#>  [150,] 73.84039 87.40166
#>  [151,] 78.38696 85.67946
#>  [152,] 75.24762 89.78652
#>  [153,] 74.91193 88.40160
#>  [154,] 73.23428 91.70418
#>  [155,] 77.13625 84.12325
#>  [156,] 74.98412 88.43696
#>  [157,] 74.58563 91.81348
#>  [158,] 76.77632 91.79018
#>  [159,] 73.80140 91.98938
#>  [160,] 73.38830 91.10504
#>  [161,] 76.43907 91.88796
#>  [162,] 77.10722 89.53162
#>  [163,] 74.53976 87.39682
#>  [164,] 74.98959 88.08377
#>  [165,] 75.25806 86.07255
#>  [166,] 73.83388 90.68870
#>  [167,] 75.36066 88.43369
#>  [168,] 75.65646 90.68200
#>  [169,] 75.05903 91.84639
#>  [170,] 75.68442 90.53093
#>  [171,] 75.88732 86.92681
#>  [172,] 75.95124 93.88284
#>  [173,] 74.91231 87.64530
#>  [174,] 76.08327 89.35366
#>  [175,] 74.15200 88.08055
#>  [176,] 74.29753 93.58396
#>  [177,] 74.17990 94.52316
#>  [178,] 73.27685 91.93800
#>  [179,] 74.08590 89.05134
#>  [180,] 73.13415 92.15600
#>  [181,] 77.32542 88.04509
#>  [182,] 75.25977 90.22902
#>  [183,] 74.35592 91.52101
#>  [184,] 72.61167 88.14375
#>  [185,] 74.43068 92.88965
#>  [186,] 76.23590 94.20010
#>  [187,] 74.76102 87.73527
#>  [188,] 74.03246 88.34874
#>  [189,] 72.66690 90.63595
#>  [190,] 79.05108 86.56877
#>  [191,] 74.19479 91.26570
#>  [192,] 75.18228 90.72127
#>  [193,] 77.56053 91.71974
#>  [194,] 74.11355 91.29697
#>  [195,] 75.97597 95.64733
#>  [196,] 76.04956 92.29370
#>  [197,] 75.18966 92.97112
#>  [198,] 74.00467 92.11360
#>  [199,] 76.48312 92.12733
#>  [200,] 76.65996 90.59860
#>  [201,] 73.56801 90.53164
#>  [202,] 74.39020 92.00459
#>  [203,] 74.01222 90.43368
#>  [204,] 74.22945 91.66921
#>  [205,] 72.67489 94.92103
#>  [206,] 77.99499 90.18710
#>  [207,] 73.72167 90.74058
#>  [208,] 75.09777 87.61101
#>  [209,] 73.72456 93.34947
#>  [210,] 76.54457 91.72873
#>  [211,] 72.87845 92.89706
#>  [212,] 76.57825 89.75286
#>  [213,] 77.19152 90.81000
#>  [214,] 74.93716 91.23242
#>  [215,] 73.21816 93.44325
#>  [216,] 73.82027 93.24955
#>  [217,] 73.29563 92.63072
#>  [218,] 75.11694 95.51008
#>  [219,] 74.24486 93.77308
#>  [220,] 76.14236 88.68927
#>  [221,] 75.14018 91.67390
#>  [222,] 76.59274 89.14985
#>  [223,] 76.39458 87.09620
#>  [224,] 75.61680 81.37897
#>  [225,] 75.73430 87.21717
#>  [226,] 75.54882 90.92510
#>  [227,] 75.46997 89.94993
#>  [228,] 72.07813 88.81846
#>  [229,] 73.51706 91.57437
#>  [230,] 71.48691 91.82105
#>  [231,] 73.48716 93.29322
#>  [232,] 75.38683 89.96235
#>  [233,] 76.06080 92.14672
#>  [234,] 74.13196 92.62700
#>  [235,] 74.52931 93.30880
#>  [236,] 75.15854 89.64933
#>  [237,] 73.94981 89.69445
#>  [238,] 73.34752 90.37839
#>  [239,] 74.91885 88.66708
#>  [240,] 73.70096 90.61868
#>  [241,] 74.19192 89.21475
#>  [242,] 74.41865 89.42067
#>  [243,] 71.42661 90.17033
#>  [244,] 75.14910 91.16880
#>  [245,] 77.33763 88.99927
#>  [246,] 72.91404 91.02371
#>  [247,] 74.34130 88.43378
#>  [248,] 74.05869 91.02166
#>  [249,] 74.38764 94.76989
#>  [250,] 74.63840 93.65415
#>  [251,] 71.97297 92.19415
#>  [252,] 73.03057 93.81402
#>  [253,] 73.43014 93.35741
#>  [254,] 76.88408 90.18619
#>  [255,] 73.03079 92.82173
#>  [256,] 75.01646 89.40659
#>  [257,] 75.01154 86.04651
#>  [258,] 74.02783 89.13187
#>  [259,] 72.82358 90.47836
#>  [260,] 74.42438 88.42535
#>  [261,] 75.22075 88.19772
#>  [262,] 73.91056 91.33838
#>  [263,] 72.64202 95.33515
#>  [264,] 75.22140 86.98830
#>  [265,] 72.66489 90.29157
#>  [266,] 75.24041 90.93650
#>  [267,] 75.77292 94.07024
#>  [268,] 75.78563 88.65944
#>  [269,] 75.65405 91.52047
#>  [270,] 72.24043 93.53042
#>  [271,] 76.19067 88.72018
#>  [272,] 74.66567 91.19459
#>  [273,] 74.65289 89.53979
#>  [274,] 73.88012 91.38221
#>  [275,] 75.63479 88.52196
#>  [276,] 74.25167 94.64930
#>  [277,] 73.18290 86.64631
#>  [278,] 75.23688 90.84892
#>  [279,] 74.16977 92.92920
#>  [280,] 74.24797 92.94250
#>  [281,] 72.63614 92.74072
#>  [282,] 72.17930 91.77439
#>  [283,] 73.96897 88.19778
#>  [284,] 74.84873 91.97253
#>  [285,] 74.10463 90.26309
#>  [286,] 75.40212 92.61224
#>  [287,] 73.28932 90.17118
#>  [288,] 72.78865 92.89818
#>  [289,] 72.41591 89.89070
#>  [290,] 75.18313 88.41837
#>  [291,] 73.86854 86.18615
#>  [292,] 75.54237 93.79260
#>  [293,] 75.63009 90.18392
#>  [294,] 73.49809 91.78434
#>  [295,] 75.48010 90.60135
#>  [296,] 76.39705 92.35029
#>  [297,] 75.01508 89.81972
#>  [298,] 73.09902 92.17359
#>  [299,] 72.22006 92.21593
#>  [300,] 74.94136 88.95977
#>  [301,] 74.34402 88.85163
#>  [302,] 72.82751 92.68257
#>  [303,] 74.14582 87.40332
#>  [304,] 74.06817 88.68301
#>  [305,] 76.93145 83.77359
#>  [306,] 74.72587 87.18190
#>  [307,] 73.26397 91.03954
#>  [308,] 74.18937 94.62396
#>  [309,] 70.91849 93.67230
#>  [310,] 72.05193 92.90217
#>  [311,] 74.68395 90.66776
#>  [312,] 73.53312 90.45383
#>  [313,] 75.52448 87.09772
#>  [314,] 75.37728 90.06408
#>  [315,] 75.62813 89.21930
#>  [316,] 76.04300 90.62079
#>  [317,] 72.40752 88.70395
#>  [318,] 74.69416 86.04263
#>  [319,] 76.55895 87.79854
#>  [320,] 73.32078 91.89085
#>  [321,] 74.22675 89.09827
#>  [322,] 74.09231 89.77972
#>  [323,] 74.15952 89.75404
#>  [324,] 74.16370 95.44682
#>  [325,] 74.19266 90.34660
#>  [326,] 73.56077 88.35933
#>  [327,] 75.13614 92.04357
#>  [328,] 73.61021 93.13266
#>  [329,] 73.03520 93.84364
#>  [330,] 74.74558 90.86613
#>  [331,] 76.47567 89.47872
#>  [332,] 75.60389 89.36079
#>  [333,] 76.28312 92.24458
#>  [334,] 73.70901 89.30868
#>  [335,] 75.25443 88.68587
#>  [336,] 76.07474 91.43617
#>  [337,] 74.29484 90.19942
#>  [338,] 76.07658 90.62602
#>  [339,] 73.67502 91.43417
#>  [340,] 74.51640 93.10622
#>  [341,] 75.78661 92.14529
#>  [342,] 74.09738 93.38876
#>  [343,] 75.65030 89.42927
#>  [344,] 74.42265 90.86397
#>  [345,] 73.72497 96.14335
#>  [346,] 75.09626 88.73826
#>  [347,] 76.23467 89.00498
#>  [348,] 74.23130 94.53417
#>  [349,] 73.56973 94.09641
#>  [350,] 74.94645 91.15650
#>  [351,] 72.75003 91.07085
#>  [352,] 74.58245 91.50210
#>  [353,] 73.77855 91.55280
#>  [354,] 74.42146 90.27020
#>  [355,] 73.50395 91.05791
#>  [356,] 73.15112 90.11005
#>  [357,] 73.39169 94.02198
#>  [358,] 74.86656 89.55508
#>  [359,] 73.87543 91.60283
#>  [360,] 75.10114 87.61886
#>  [361,] 74.30493 90.62734
#>  [362,] 75.13556 90.45938
#>  [363,] 74.29623 94.56339
#>  [364,] 76.22928 87.20951
#>  [365,] 73.17621 90.86320
#>  [366,] 74.63776 88.81683
#>  [367,] 75.17025 89.61358
#>  [368,] 78.02007 89.86220
#>  [369,] 73.43088 86.39552
#>  [370,] 77.02960 92.18024
#>  [371,] 71.30735 91.14668
#>  [372,] 76.15273 88.27182
#>  [373,] 70.64902 91.71341
#>  [374,] 76.51843 90.22526
#>  [375,] 74.67682 90.20276
#>  [376,] 76.17490 89.86911
#>  [377,] 73.16593 90.73634
#>  [378,] 72.91233 89.67382
#>  [379,] 73.30052 92.82124
#>  [380,] 71.80493 93.62824
#>  [381,] 73.98775 90.42871
#>  [382,] 73.62079 91.64083
#>  [383,] 75.62971 89.88645
#>  [384,] 73.71069 93.70901
#>  [385,] 74.80525 87.84846
#>  [386,] 74.37540 93.17145
#>  [387,] 74.24121 91.00100
#>  [388,] 73.04407 91.98775
#>  [389,] 72.13517 93.53688
#>  [390,] 75.99836 86.94488
#>  [391,] 75.43753 93.41343
#>  [392,] 76.66724 87.77252
#>  [393,] 74.16521 93.30144
#>  [394,] 75.03953 91.38509
#>  [395,] 71.19434 91.36390
#>  [396,] 74.25263 90.27966
#>  [397,] 75.20659 93.35496
#>  [398,] 75.49978 91.81161
#>  [399,] 75.29877 87.42269
#>  [400,] 72.27287 91.98207
#>  [401,] 76.33637 92.49500
#>  [402,] 73.57968 94.77645
#>  [403,] 75.09954 89.64637
#>  [404,] 74.24748 88.79953
#>  [405,] 75.30097 86.09875
#>  [406,] 74.77391 90.69847
#>  [407,] 74.53783 90.99575
#>  [408,] 75.93186 92.54236
#>  [409,] 73.44621 90.51159
#>  [410,] 76.85568 90.56770
#>  [411,] 73.28348 90.24015
#>  [412,] 73.06030 94.18308
#>  [413,] 76.50898 96.04227
#>  [414,] 75.70137 98.25166
#>  [415,] 75.08737 90.97307
#>  [416,] 75.71842 90.76962
#>  [417,] 71.92817 91.86209
#>  [418,] 76.23415 88.28586
#>  [419,] 73.58230 89.49364
#>  [420,] 74.76719 88.22821
#>  [421,] 73.18155 94.50912
#>  [422,] 73.27814 88.67677
#>  [423,] 74.15007 93.17353
#>  [424,] 73.58559 90.41061
#>  [425,] 75.11258 89.67268
#>  [426,] 74.97507 87.00728
#>  [427,] 72.27832 90.57579
#>  [428,] 73.52097 91.37999
#>  [429,] 74.79709 91.21248
#>  [430,] 72.72125 90.69632
#>  [431,] 72.86717 88.80455
#>  [432,] 72.86080 93.68343
#>  [433,] 74.51528 90.72712
#>  [434,] 74.36085 90.23285
#>  [435,] 74.64397 87.78755
#>  [436,] 76.15852 90.35450
#>  [437,] 75.31333 87.77175
#>  [438,] 73.56186 88.94542
#>  [439,] 75.84536 85.39693
#>  [440,] 75.29899 90.30939
#>  [441,] 73.70331 89.23951
#>  [442,] 72.91595 92.87803
#>  [443,] 73.23944 87.99097
#>  [444,] 69.17686 94.53372
#>  [445,] 74.62320 92.98164
#>  [446,] 74.59354 90.53958
#>  [447,] 73.67072 94.62485
#>  [448,] 72.09691 90.83091
#>  [449,] 73.82209 89.71572
#>  [450,] 77.36084 88.79574
#>  [451,] 72.40145 98.54164
#>  [452,] 72.50551 89.70367
#>  [453,] 74.53396 89.48489
#>  [454,] 74.92743 91.17762
#>  [455,] 73.57958 93.67765
#>  [456,] 76.86837 90.49632
#>  [457,] 75.73756 87.37355
#>  [458,] 72.45199 96.43391
#>  [459,] 73.14974 91.26986
#>  [460,] 75.47548 95.37103
#>  [461,] 72.89239 87.93241
#>  [462,] 72.96832 89.01649
#>  [463,] 73.95746 92.32306
#>  [464,] 77.91691 82.71401
#>  [465,] 73.27208 89.85837
#>  [466,] 77.10610 91.98673
#>  [467,] 75.29117 91.01205
#>  [468,] 74.20005 90.54741
#>  [469,] 75.41166 91.26653
#>  [470,] 74.86493 91.67625
#>  [471,] 76.82478 85.85993
#>  [472,] 76.87564 92.97553
#>  [473,] 75.15214 90.96770
#>  [474,] 75.55798 89.51596
#>  [475,] 73.21691 93.37867
#>  [476,] 76.88377 89.08686
#>  [477,] 74.30486 92.09242
#>  [478,] 74.03042 87.40172
#>  [479,] 78.12686 88.67933
#>  [480,] 74.29267 87.78784
#>  [481,] 71.80198 91.86177
#>  [482,] 74.37991 92.70084
#>  [483,] 74.40721 91.23958
#>  [484,] 72.00721 92.57949
#>  [485,] 74.90982 90.71246
#>  [486,] 75.24807 91.85986
#>  [487,] 74.88232 88.66073
#>  [488,] 75.72296 89.14346
#>  [489,] 73.21615 91.28751
#>  [490,] 73.32140 93.00765
#>  [491,] 75.89659 88.28328
#>  [492,] 76.37342 85.12024
#>  [493,] 74.52795 93.63118
#>  [494,] 74.11276 88.85176
#>  [495,] 76.24759 89.12625
#>  [496,] 69.44231 89.71288
#>  [497,] 75.66257 92.85319
#>  [498,] 76.07132 91.22113
#>  [499,] 73.20912 91.75907
#>  [500,] 75.12419 88.81001
#>  [ reached getOption("max.print") -- omitted 500 rows ]