breastcancer {LearnBayes}	R Documentation

Survival experience of women with breast cancer under treatment

Description

Collett (1994) describes a study to evaluate the effectiveness of a histochemical marker in predicting the survival experience of women with breast cancer.

Usage

breastcancer

Format

A data frame with 45 observations on the following 3 variables.

time

survival time in months

status

censoring indicator where 1 (0) indicates a complete (censored) survival time

stain

indicates by a 0 (1) if tumor was negatively (positively) stained

Source

Collett, D. (1994), Modelling Survival Data in Medical Research, London: Chapman and Hall.

[Package LearnBayes version 2.15.1 Index]