Setup for Career Trajectory Application

Setups the data matrices for the use of WinBUGS in the career trajectory application.

careertraj.setup(data)

Arguments

data
data matrix for ballplayers with variables Player, Year, Age, G, AB, R, H, X2B, X3B, HR, RBI, BB, SO

Value

player.names

vector of player names

y

matrix of home runs for players where a row corresponds to the home runs for a player during all the years of his career

n

matrix of AB-SO for all players

x

matrix of ages for all players for all years of their careers

T

vector of number of seasons for all players

N

number of players

Examples

data(sluggerdata)
careertraj.setup(sluggerdata)

#> $player.names
#>  [1] "Aaron"     "Greenberg" "Killebrew" "Mantle"    "Mays"      "McCovey"  
#>  [7] "Ott"       "Ruth"      "Schmidt"   "Sosa"     
#> 
#> $y
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   13   27   26   44   30   39   40   34   45    44    24    32    44
#>  [2,]    0   12   26   36    1   40   58   33   41     2    13    44    25
#>  [3,]    0    4    5    2    0   42   31   46   48    45    49    25    39
#>  [4,]   13   23   21   27   37   52   34   42   31    40    54    30    15
#>  [5,]   20    4   41   51   36   35   29   34   29    40    49    38    47
#>  [6,]   13   13   18   20   44   18   39   36   31    36    45    39    18
#>  [7,]    0    1   18   42   25   29   38   23   35    31    33    31    36
#>  [8,]    0    4    3    2   11   29   54   59   35    41    46    25    47
#>  [9,]    1   18   36   38   38   38   21   45   48    31    35    40    36
#> [10,]    4   15   10    8   33   25   36   40   36    66    63    50    64
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23]
#>  [1,]    39    29    44    38    47    34    40    20    12    10
#>  [2,]     0     0     0     0     0     0     0     0     0     0
#>  [3,]    44    17    49    41    28    26     5    13    14     0
#>  [4,]    35    19    23    22    18     0     0     0     0     0
#>  [5,]    52    37    22    23    13    28    18     8     6     0
#>  [6,]    14    29    22    23     7    28    12    15     1     0
#>  [7,]    27    19    27    30    18    26    21     1     0     0
#>  [8,]    60    54    46    49    46    41    34    22     6     0
#>  [9,]    33    37    35    12     6     0     0     0     0     0
#> [10,]    49    40    35    14     0     0     0     0     0     0
#> 
#> $n
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  429  541  555  557  552  575  527  539  519   537   524   489   507
#>  [2,]    1  371  500  528   40  493  464  405  498    55   230   435   329
#>  [3,]   10   49   60   23   19  430  336  432  410   410   442   332   471
#>  [4,]  267  438  371  436  420  434  399  399  415   402   402   299   140
#>  [5,]  404  110  508  520  513  523  544  517  525   495   536   513   506
#>  [6,]  157  207  268  194  445  291  422  402  346   452   425   420   272
#>  [7,]   51  154  399  507  486  453  527  532  539   535   493   476   480
#>  [8,]    6   69  113  105  259  374  377  459  326   429   448   291   419
#>  [9,]   19  231  430  382  435  422  410  426  429   283   383   386   412
#> [10,]  136  382  218  199  463  334  430  364  468   472   454   436   424
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23]
#>  [1,]   503   544   500   453   437   394   341   311   414   233
#>  [2,]     0     0     0     0     0     0     0     0     0     0
#>  [3,]   436   225   471   443   404   342   189   272   242     0
#>  [4,]   363   285   257   327   338     0     0     0     0     0
#>  [5,]   487   471   394   417   332   388   294   196   162     0
#>  [6,]   218   305   268   333   183   372   294   283    90     0
#>  [7,]   346   486   457   488   332   352   410    53     4     0
#>  [8,]   451   449   439   457   483   395   369   302    48     0
#>  [9,]   432   468   442   348   131     0     0     0     0     0
#> [10,]   412   374   345   296     0     0     0     0     0     0
#> 
#> $x
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   20   21   22   23   24   25   26   27   28    29    30    31    32
#>  [2,]   19   22   23   24   25   26   27   28   29    30    34    35    36
#>  [3,]   18   19   20   21   22   23   24   25   26    27    28    29    30
#>  [4,]   19   20   21   22   23   24   25   26   27    28    29    30    31
#>  [5,]   20   21   23   24   25   26   27   28   29    30    31    32    33
#>  [6,]   21   22   23   24   25   26   27   28   29    30    31    32    33
#>  [7,]   17   18   19   20   21   22   23   24   25    26    27    28    29
#>  [8,]   19   20   21   22   23   24   25   26   27    28    29    30    31
#>  [9,]   22   23   24   25   26   27   28   29   30    31    32    33    34
#> [10,]   20   21   22   23   24   25   26   27   28    29    30    31    32
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23]
#>  [1,]    33    34    35    36    37    38    39    40    41    42
#>  [2,]     0     0     0     0     0     0     0     0     0     0
#>  [3,]    31    32    33    34    35    36    37    38    39     0
#>  [4,]    32    33    34    35    36     0     0     0     0     0
#>  [5,]    34    35    36    37    38    39    40    41    42     0
#>  [6,]    34    35    36    37    38    39    40    41    42     0
#>  [7,]    30    31    32    33    34    35    36    37    38     0
#>  [8,]    32    33    34    35    36    37    38    39    40     0
#>  [9,]    35    36    37    38    39     0     0     0     0     0
#> [10,]    33    34    35    36     0     0     0     0     0     0
#> 
#> $T
#>  [1] 23 13 22 18 22 22 22 22 18 17
#> 
#> $N
#> [1] 10
#>