chemotherapy {LearnBayes}	R Documentation

Chemotherapy treatment effects on ovarian cancer

Description

Edmunson et al (1979) studied the effect of different chemotherapy treatments following surgical treatment of ovarian cancer.

Usage

chemotherapy

Format

A data frame with 26 observations on the following 5 variables.

patient

patient number

time

survival time in days following treatment

status

indicates if time is censored (0) or actually observed (1)

treat

control group (0) or treatment group (1)

age

age of the patient

Source

Edmonson, J., Felming, T., Decker, D., Malkasian, G., Jorgensen, E., Jefferies, J.,Webb, M., and Kvols, L. (1979), Different chemotherapeutic sensitivities and host factors affecting prognosis in advanced ovarian carcinoma versus minimal residual disease, Cancer Treatment Reports, 63, 241-247.

[Package LearnBayes version 2.15.1 Index]