election {LearnBayes}	R Documentation

Florida election data

Description

For each of the Florida counties in the 2000 presidential election, the number of votes for George Bush, Al Gore, and Pat Buchanan is recorded. Also the number of votes for the minority candidate Ross Perot in the 1996 presidential election is recorded.

Usage

election

Format

A data frame with 67 observations on the following 5 variables.

county

name of Florida county

perot

number of votes for Ross Perot in 1996 election

gore

number of votes for Al Gore in 2000 election

bush

number of votes for George Bush in 2000 election

buchanan

number of votes for Pat Buchanan in 2000 election

[Package LearnBayes version 2.15.1 Index]