hearttransplants {LearnBayes}	R Documentation

Heart transplant mortality data

Description

The number of deaths within 30 days of heart transplant surgery for 94 U.S. hospitals that performed at least 10 heart transplant surgeries. Also the exposure, the expected number of deaths, is recorded for each hospital.

Usage

hearttransplants

Format

A data frame with 94 observations on the following 2 variables.

e

expected number of deaths (the exposure)

y

observed number of deaths within 30 days of heart transplant surgery

Source

Christiansen, C. and Morris, C. (1995), Fitting and checking a two-level Poisson model: modeling patient mortality rates in heart transplant patients, in Berry, D. and Stangl, D., eds, Bayesian Biostatistics, Marcel Dekker.

[Package LearnBayes version 2.15.1 Index]