indepmetrop {LearnBayes}	R Documentation

Independence Metropolis independence chain of a posterior distribution

Description

Simulates iterates of an independence Metropolis chain with a normal proposal density for an arbitrary real-valued posterior density defined by the user

Usage

indepmetrop(logpost,proposal,start,m,...)

Arguments

logpost
function defining the log posterior density

proposal
a list containing mu, an estimated mean and var, an estimated variance-covariance matrix, of the normal proposal density

start
vector containing the starting value of the parameter

m
the number of iterations of the chain

...
data that is used in the function logpost

Value

par
a matrix of simulated values where each row corresponds to a value of the vector parameter

accept
the acceptance rate of the algorithm

Author(s)

Jim Albert

Examples

data=c(6,2,3,10)
proposal=list(mu=array(c(2.3,-.1),c(2,1)),var=diag(c(1,1)))
start=array(c(0,0),c(1,2))
m=1000
fit=indepmetrop(logctablepost,proposal,start,m,data)

[Package LearnBayes version 2.15.1 Index]