iowagpa {LearnBayes}	R Documentation

Admissions data for an university

Description

Students at a major university are categorized with respect to their high school rank and their ACT score. For each combination of high school rank and ACT score, one records the mean grade point average (GPA).

Usage

iowagpa

Format

A data frame with 40 observations on the following 4 variables.

gpa

mean grade point average

n

sample size

HSR

high school rank

ACT

act score

Source

Albert, J. (1994), A Bayesian approach to estimation of GPA's of University of Iowa freshmen under order restrictions, Journal of Educational Statistics, 19, 1-22.

[Package LearnBayes version 2.15.1 Index]