logctablepost {LearnBayes}	R Documentation

Log posterior of difference and sum of logits in a 2x2 table

Description

Computes the log posterior density for the difference and sum of logits in a 2x2 contingency table for independent binomial samples and uniform prior placed on the logits

Usage

logctablepost(theta,data)

Arguments

theta
vector of parameter values "difference of logits" and "sum of logits")

data
vector containing number of successes and failures for first sample, and then second sample

Value

value of the log posterior

Author(s)

Jim Albert

Examples

s1=6; f1=2; s2=3; f2=10
data=c(s1,f1,s2,f2)
theta=c(2,4)
logctablepost(theta,data)

[Package LearnBayes version 2.15.1 Index]