mnormt.twosided {LearnBayes}	R Documentation

Bayesian test of a two-sided hypothesis about a normal mean

Description

Bayesian test that a normal mean is equal to a specified value using a normal prior

Usage

mnormt.twosided(m0, prob, t, data)

Arguments

m0
value of the mean to be tested

prob
prior probability of the hypothesis

t
vector of values of the prior standard deviation under the alternative hypothesis

data
vector containing the sample mean, the sample size, and the known value of the population standard deviation

Value

bf
vector of values of the Bayes factor in support of the null hypothesis

post
vector of posterior probabilities of the null hypothesis

Author(s)

Jim Albert

Examples

m0=170
prob=.5
tau=c(.5,1,2,4,8)
samplesize=10
samplemean=176
popsd=3
data=c(samplemean,samplesize,popsd)
mnormt.twosided(m0,prob,tau,data)

[Package LearnBayes version 2.15.1 Index]