normal.select {LearnBayes}	R Documentation

Selection of Normal Prior Given Knowledge of Two Quantiles

Description

Finds the mean and standard deviation of a normal density that matches knowledge of two quantiles of the distribution.

Usage

normal.select(quantile1, quantile2)

Arguments

quantile1
list with components p, the value of the first probability, and x, the value of the first quantile

quantile2
list with components p, the value of the second probability, and x, the value of the second quantile

Value

mean
mean of the matching normal distribution

sigma
standard deviation of the matching normal distribution

Author(s)

Jim Albert

Examples

# person believes the 15th percentile of the prior is 100
# and the 70th percentile of the prior is 150
quantile1=list(p=.15,x=100)
quantile2=list(p=.7,x=150)
normal.select(quantile1,quantile2)

[Package LearnBayes version 2.15.1 Index]