normchi2post {LearnBayes}	R Documentation

Log posterior density for mean and variance for normal sampling

Description

Computes the log of the posterior density of a mean M and a variance S2 when a sample is taken from a normal density and a standard noninformative prior is used.

Usage

normchi2post(theta,data)

Arguments

theta
vector of parameter values M and S2

data
vector containing the sample observations

Value

value of the log posterior

Author(s)

Jim Albert

Examples

parameter=c(25,5)
data=c(20, 32, 21, 43, 33, 21, 32)
normchi2post(parameter,data)

[Package LearnBayes version 2.15.1 Index]