normpostpred {LearnBayes}	R Documentation

Posterior predictive simulation from Bayesian normal sampling model

Description

Given simulated draws from the posterior from a normal sampling model, outputs simulated draws from the posterior predictive distribution of a statistic of interest.

Usage

normpostpred(parameters,sample.size,f=min)

Arguments

parameters
list of simulated draws from the posterior where mu contains the normal mean and sigma2 contains the normal variance

sample.size
size of sample of future sample

f
function defining the statistic

Value

simulated sample of the posterior predictive distribution of the statistic

Author(s)

Jim Albert

Examples

# finds posterior predictive distribution of the min statistic of a future sample of size 15
data(darwin)
s=normpostsim(darwin$difference)
sample.size=15
sim.stats=normpostpred(s,sample.size,min)

[Package LearnBayes version 2.15.1 Index]