poissgamexch {LearnBayes}	R Documentation

Log posterior of Poisson/gamma exchangeable model

Description

Computes the log posterior density of log alpha and log mu for a Poisson/gamma exchangeable model

Usage

poissgamexch(theta,datapar)

Arguments

theta
vector of parameter values of log alpha and log mu

datapar
list with components data, a matrix with columns e and y, and z0, prior hyperparameter

Value

value of the log posterior

Author(s)

Jim Albert

Examples

e=c(532,584,672,722,904)
y=c(0,0,2,1,1)
data=cbind(e,y)
theta=c(-4,0)
z0=.5
datapar=list(data=data,z0=z0)
poissgamexch(theta,datapar)

[Package LearnBayes version 2.15.1 Index]