predplot {LearnBayes}	R Documentation

Plot of predictive distribution for binomial sampling with a beta prior

Description

For a proportion problem with a beta prior, plots the prior predictive distribution of the number of successes in n trials and displays the observed number of successes.

Usage

predplot(prior,n,yobs)

Arguments

prior
vector of parameters for beta prior

n
sample size

yobs
observed number of successes

Author(s)

Jim Albert

Examples

prior=c(3,10)  # proportion has a beta(3, 10) prior
n=20   # sample size
yobs=10  # observed number of successes
predplot(prior,n,yobs)

[Package LearnBayes version 2.15.1 Index]