Gibbs sampling for a robust regression model

Implements Gibbs sampling for a robust t sampling model with location mu, scale sigma, and degrees of freedom v

robustt(y,v,m)

Arguments

y
vector of data values

v
degrees of freedom for t model

m
the number of cycles of the Gibbs sampler

Value

mu

vector of simulated values of mu

s2

vector of simulated values of sigma2

lam

matrix of simulated draws of lambda, where each row corresponds to a single draw

Examples

data=c(-67,-48,6,8,14,16,23,24,28,29,41,49,67,60,75)
fit=robustt(data,4,1000)