triplot {LearnBayes}	R Documentation

Plot of prior, likelihood and posterior for a proportion

Description

For a proportion problem with a beta prior, plots the prior, likelihood and posterior on one graph.

Usage

triplot(prior,data,where="topright")

Arguments

prior
vector of parameters for beta prior

data
vector consisting of number of successes and number of failures

where
the location of the legend for the plot

Author(s)

Jim Albert

Examples

prior=c(3,10)  # proportion has a beta(3, 10) prior
data=c(10,6)   # observe 10 successes and 6 failures
triplot(prior,data)

[Package LearnBayes version 2.15.1 Index]